virtualEnv/coinflip/requirements.txt
Django
mysqlclient
mysql-connector-python